+++ docs/changes.rst
@@ -18,6 +18,7 @@ Weblate 5.17
 
 .. rubric:: Bug fixes
 
+* Removing the last team from a project token is no longer allowed to prevent the token from becoming invisible in the project access management.
 * Fixed sorting by the **Unreviewed** column in listings.
 * Fixed false positive in :ref:`check-xml-chars-around-tags` for Arabic letter Waw ("و") adjacent to XML tags.
 * :ref:`addon-weblate.git.squash` better handle commits applied upstream.